Accessibility for button
Added accessibility for buttons on welcome page, map page, and job details page, where the talk back will read helpful text when button is focused on. e.g. the screen reader will read " press to take a picture button". which was done by adding automation properties. is in accessible set to true with the helptext property where the text is added of what user would want to hear when the button is focued.